backend/requirements.txt
spotipy
pandas
sqlalchemy
python-dotenv
psycopg2-binary
flask
flask-cors